Bröt upp class5 i Personal klasser.
