+++ PROGRAM_MASTER.py
@@ -1,5 +1,11 @@
+from class5 import *
+
+
 def main():
     print("Här ska vi köra programmet")
+
+    menu = Menu()
+    menu.run()
 
 
 if __name__ == '__main__':
